elasticsearch/requirements.txt
pika==1.3.2
python-dotenv==0.21.0
elasticsearch==7.13.4